feat(dedicated): server network interfaces page improvements
ref: MANAGER-12223
