+++ packages/manager/apps/dedicated/client/app/dedicated/server/dashboard/dashboard.component.js
@@ -7,6 +7,7 @@ export default {
     biosSettings: '<',
     bringYourOwnImage: '<',
     changeOwnerAvailable: '<',
+    currentActiveLink: '<',
     isMonitoringOptionsAvailable: '<',
     changeOwnerUrl: '<',
     eligibleData: '<',
