+++ packages/manager/apps/dedicated/client/app/dedicated/server/dashboard/dashboard.controller.js
@@ -1,7 +1,4 @@
 import get from 'lodash/get';
-import head from 'lodash/head';
-import map from 'lodash/map';
-import some from 'lodash/some';
 
 import {
   MONITORING_STATUSES,
@@ -10,7 +7,6 @@ import {
   WEATHERMAP_URL,
   COMMIT_IMPRESSION_TRACKING_DATA,
   RECOMMIT_IMPRESSION_TRACKING_DATA,
-  HIDE_MRTG_FOR_SERVER_RANGES,
   VMS_URL_OTHERS,
 } from './dashboard.constants';
 import { NEW_RANGE, OPERATING_SYSTEM_ENUM } from '../details/server.constants';
@@ -70,28 +66,6 @@ export default class DedicatedServerDashboard {
     };
 
     this.pattern = NEW_RANGE.PATTERN;
-
-    this.serverStatsLoad = {
-      loading: true,
-      error: false,
-    };
-
-    this.serverStats = {
-      model: null,
-      modelconst: null,
-      networkChoice: null,
-      periodeChoice: null,
-      typeChoice: null,
-      mrtgShow: false,
-    };
-
-    this.networks = [];
-
-    this.$scope.$on('reloadChart', () => {
-      this.getStatistics();
-    });
-
-    this.loadStatistics();
   }
 
   getVmsLink() {
@@ -101,163 +75,8 @@ export default class DedicatedServerDashboard {
     );
   }
 
-  createChart(data) {
-    this.series = [];
-    this.data = [];
-    this.labels = map(get(data, 'download.values'), (value) =>
-      moment.unix(value.timestamp).calendar(),
-    );
-    this.series.push(
-      this.$translate.instant('server_tab_STATS_legend_download'),
-    );
-    this.series.push(this.$translate.instant('server_tab_STATS_legend_upload'));
-    this.data.push(
-      DedicatedServerDashboard.convertData(get(data, 'download.values')),
-    );
-    this.data.push(
-      DedicatedServerDashboard.convertData(get(data, 'upload.values')),
-    );
-
-    const { unit } = head(get(data, 'download.values'));
-    const yLabel =
-      unit === 'bps'
-        ? this.$translate.instant(
-            'server_configuration_mitigation_statistics_unit_KB',
-          )
-        : unit;
-    this.options = {
-      scales: {
-        yAxes: [
-          {
-            scaleLabel: {
-              display: true,
-              labelString: yLabel,
-            },
-          },
-        ],
-        xAxes: [
-          {
-            scaleLabel: {
-              display: true,
-              labelString: this.$translate.instant(
-                'server_usage_chart_xaxis_label',
-              ),
-            },
-          },
-        ],
-      },
-    };
-  }
-
-  loadStatistics() {
-    const nameServer = null;
-    this.serverStatsLoad.loading = true;
-    this.serverStatsLoad.error = false;
-    const promises = {
-      interfaces: this.Server.getNetworkInterfaces(this.$stateParams.productId),
-      statConst: this.Server.getStatisticsConst(),
-    };
-
-    this.$q
-      .all(promises)
-      .then((stats) => {
-        const hasVrack = some(
-          stats.interfaces,
-          (networkInterface) => networkInterface.linkType === 'private',
-        );
-        if (!hasVrack) {
-          stats.interfaces.push({
-            linkType: 'no_vrack',
-            disabled: true,
-          });
-        }
-
-        this.networks = map(stats.interfaces, (networkInterface) => ({
-          id: networkInterface.mac,
-          linkType: networkInterface.linkType,
-          displayName: this.$translate.instant(
-            `server_tab_stats_network_${networkInterface.linkType}`,
-            {
-              t0: networkInterface.mac,
-            },
-          ),
-          disabled: networkInterface.disabled,
-        }));
-
-        this.serverStats.networkChoice = this.networks[0].id;
-
-        this.serverStats.modelconst = stats.statConst;
-
-        if (this.serverStats.periodeChoice === null) {
-          this.serverStats.periodeChoice = stats.statConst.defaultPeriod;
-        }
-
-        if (this.serverStats.typeChoice === null) {
-          this.serverStats.typeChoice = stats.statConst.defaultType;
-        }
-
-        return this.Server.getStatistics(
-          this.$stateParams.productId,
-          this.serverStats.networkChoice,
-          this.serverStats.typeChoice,
-          this.serverStats.periodeChoice,
-        );
-      })
-      .then((statistics) => {
-        this.serverStats.model = statistics;
-        this.createChart(statistics);
-      })
-      .catch((data) => {
-        this.serverStatsLoad.error = true;
-
-        if (nameServer && data) {
-          this.dedicatedServer.setMessage(
-            this.$translate.instant('server_tab_STATS_loading_fail'),
-            {
-              ...data.data,
-              typpe: 'ERROR',
-            },
-          );
-        }
-      })
-      .finally(() => {
-        this.serverStatsLoad.loading = false;
-      });
-  }
-
   rebootServer() {
     return this.$state.go('app.dedicated-server.server.dashboard.reboot');
-  }
-
-  getStatistics() {
-    this.serverStatsLoad.loading = true;
-    this.serverStatsLoad.error = false;
-
-    this.Server.getStatistics(
-      this.$stateParams.productId,
-      this.serverStats.networkChoice,
-      this.serverStats.typeChoice,
-      this.serverStats.periodeChoice,
-    )
-      .then((statistics) => {
-        this.serverStats.model = statistics;
-        this.createChart(statistics);
-      })
-      .catch((data) => {
-        this.serverStatsLoad.error = true;
-        if (data) {
-          this.dedicatedServer.setMessage(
-            this.$translate.instant('server_tab_STATS_loading_fail'),
-            {
-              ...data.data,
-              type: 'ERROR',
-            },
-          );
-        }
-      })
-      .finally(() => {
-        this.serverStatsLoad.loading = false;
-      });
   }
 
   getMonitoringStatus() {
@@ -319,12 +138,6 @@ export default class DedicatedServerDashboard {
           },
         );
       });
-  }
-
-  static convertData(list) {
-    return map(list, (value) =>
-      value.unit === 'bps' ? (value.y / 1024).toFixed(2) : value.y,
-    );
   }
 
   onBillingInformationError(error) {
@@ -414,12 +227,6 @@ export default class DedicatedServerDashboard {
     });
 
     return this.goToMonitoringUpdate();
-  }
-
-  displayMRTG() {
-    return !HIDE_MRTG_FOR_SERVER_RANGES.includes(
-      this.technicalDetails?.server?.range,
-    );
   }
 
   getOperatingSystemLabel() {
