+++ packages/manager/apps/dedicated/client/app/dedicated/server/dashboard/dashboard.html
@@ -30,6 +30,11 @@
         </span>
     </oui-message>
 
+    <!-- QUOTA -->
+    <div class="mb-5" data-ng-if="$ctrl.trafficInformations.traffic.data.hasQuota"
+        data-ng-include="'dedicated/server/consumption/consumption.html'">
+    </div>
+
     <div class="row d-md-flex mb-5">
         <div class="col-md-4 mb-3">
             <!-- GENERAL INFORMATIONS -->
@@ -844,144 +849,4 @@
         </div>
     </div>
 
-    <!-- QUOTA -->
-    <div
-        class="mb-5"
-        data-ng-if="$ctrl.trafficInformations.traffic.data.hasQuota"
-        data-ng-include="'dedicated/server/consumption/consumption.html'"
-    ></div>
-
-    <!-- MRTG -->
-    <div
-        class="oui-tile mb-5"
-        data-ng-if="!$ctrl.server.isExpired && $ctrl.displayMRTG()"
-    >
-        <h4
-            class="oui-tile__title oui-heading_4"
-            data-translate="server_tab_stats_legend_mrtg"
-        ></h4>
-        <p
-            class="oui-tile__legend"
-            data-translate="server_tab_stats_title_info"
-        ></p>
-        <div class="oui-tile__body">
-            <div
-                class="oui-tile__item text-center"
-                data-ng-if="$ctrl.serverStatsLoad.loading"
-            >
-                <oui-spinner></oui-spinner>
-            </div>
-            <div
-                class="oui-tile__item text-center"
-                data-ng-if="!$ctrl.serverStatsLoad.loading"
-            >
-                <div class="oui-tile__definition">
-                    <div class="oui-box mb-5">
-                        <div
-                            class="d-md-flex align-items-md-center justify-content-md-between"
-                        >
-                            <div>
-                                <label
-                                    for="networkChoice"
-                                    class="oui-label d-inline-block align-middle mr-2"
-                                    data-translate="server_tab_stats_network"
-                                >
-                                </label>
-                                <div
-                                    class="oui-select mb-0 oui-select_inline align-middle"
-                                >
-                                    <select
-                                        class="oui-select__input"
-                                        id="networkChoice"
-                                        name="datatype"
-                                        data-ng-model="$ctrl.serverStats.networkChoice"
-                                        data-ng-change="$ctrl.getStatistics()"
-                                    >
-                                        <option
-                                            data-ng-repeat="network in $ctrl.networks | orderBy:'displayName' track by network.id"
-                                            data-ng-value="network.id"
-                                            data-ng-disabled="network.disabled"
-                                            data-ng-selected="{{ network.id === $ctrl.serverStats.networkChoice }}"
-                                            data-ng-bind="network.displayName"
-                                        >
-                                        </option>
-                                    </select>
-                                    <span
-                                        class="oui-icon oui-icon-chevron-down"
-                                        aria-hidden="true"
-                                    ></span>
-                                </div>
-                            </div>
-                            <div class="oui-select">
-                                <label
-                                    for="typeChoice"
-                                    class="oui-label d-inline-block align-middle mr-2"
-                                    data-translate="server_tab_stats_type"
-                                >
-                                </label>
-                                <div
-                                    class="oui-select mb-0 oui-select_inline align-middle"
-                                >
-                                    <select
-                                        class="oui-select__input"
-                                        id="typeChoice"
-                                        name="datatype"
-                                        data-ng-options="('server_tab_stats_type_' + type | translate) for type in $ctrl.serverStats.modelconst.types"
-                                        data-ng-model="$ctrl.serverStats.typeChoice"
-                                        data-ng-change="$ctrl.getStatistics()"
-                                    >
-                                    </select>
-                                    <span
-                                        class="oui-icon oui-icon-chevron-down"
-                                        aria-hidden="true"
-                                    ></span>
-                                </div>
-                            </div>
-                            <div class="oui-select">
-                                <label
-                                    for="periodeChoice"
-                                    class="oui-label d-inline-block align-middle mr-2"
-                                    data-translate="server_tab_stats_period"
-                                >
-                                </label>
-                                <div
-                                    class="oui-select mb-0 oui-select_inline align-middle"
-                                >
-                                    <select
-                                        class="oui-select__input"
-                                        id="periodeChoice"
-                                        name="datatype"
-                                        data-ng-options="('server_tab_stats_period_' + periode | translate) for periode in $ctrl.serverStats.modelconst.periods"
-                                        data-ng-model="$ctrl.serverStats.periodeChoice"
-                                        data-ng-change="$ctrl.getStatistics()"
-                                    >
-                                    </select>
-                                    <span
-                                        class="oui-icon oui-icon-chevron-down"
-                                        aria-hidden="true"
-                                    ></span>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <canvas
-                        id="mrtgStatsChart"
-                        class="chart chart-line"
-                        data-ng-if="!$ctrl.serverStatsLoad.error"
-                        data-chart-labels="$ctrl.labels"
-                        data-chart-series="$ctrl.series"
-                        data-chart-data="$ctrl.data"
-                        data-chart-options="$ctrl.options"
-                    >
-                    </canvas>
-                    <div
-                        class="alert alert-danger"
-                        role="alert"
-                        data-ng-if="$ctrl.serverStatsLoad.error"
-                        data-translate="server_tab_stats_no_chart"
-                    ></div>
-                </div>
-            </div>
-        </div>
-    </div>
 </div>
