+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/translations/Messages_fr_FR.json
@@ -75,9 +75,8 @@
   "server_tab_STATS_legend_y_TRAFFIC": "Niveau de traffic",
   "server_tab_STATS_legend_no_data": "Pas de données",
   "server_tab_stats_no_chart": "Impossible d'afficher les données",
-  "server_tab_stats_title": "Graphiques d'utilisation réseaux - MRTG",
-  "server_tab_stats_title_info": "Multi Router Traffic Grapher - logiciel permettant de créer des graphiques sur le trafic réseau",
-  "server_tab_stats_legend_mrtg": "MRTG",
+  "server_tab_stats_title_info": "Multi Router Traffic Grapher (MRTG) est un logiciel permettant de créer des graphiques sur le trafic réseau",
+  "server_tab_stats_legend_mrtg": "Monitoring Réseau",
   "server_configuration_state_title": "Statut",
   "server_configuration_state_OK": "Normal",
   "server_configuration_state_ERROR": "En erreur",
@@ -766,8 +765,9 @@
   "server_traffic_cancel_button": "Resilier",
   "server_traffic_loading_error": "Une erreur est survenue lors des chargements des informations.",
   "server_traffic_option_released": "(non renouvelé)",
-  "server_consumption_label": "Consommation",
+  "server_consumption_label": "Consommation de bande passante",
   "server_consumption_explanation": "La consommation de bande passante du serveur est limitée. Une fois la limite dépassée, la bande passante sortante du serveur sera ralentie. Pour régler ce problème, augmentez votre quota de bande passante sortante.",
+  "server_consumption_explanation_link": "Plus de détails",
   "server_consumption_summary_days": "Reste {{t0}} sur {{t1}} jours",
   "server_consumption_summary_hours": "Reste {{t0}} sur {{t1}} heures",
   "server_consumption_throttled": "Service ralenti",
