+++ packages/manager/apps/dedicated/client/app/dedicated/server/interfaces/interfaces.component.js
@@ -6,6 +6,7 @@ export default {
     bandwidthOption: '<',
     bandwidthVrackOption: '<',
     bandwidthVrackOrderOptions: '<',
+    currentActiveLink: '<',
     failoverIps: '<',
     guideUrl: '<',
     interfaces: '<',
@@ -18,6 +19,8 @@ export default {
     specifications: '<',
     resiliatePublicBandwidthLink: '<',
     resiliatePrivateBandwidthLink: '<',
+    trafficInformations: '<',
+    technicalDetails: '<',
     urls: '<',
     user: '<',
   },
