+++ packages/manager/apps/dedicated/client/app/dedicated/server/interfaces/translations/Messages_fr_FR.json
@@ -1,13 +1,9 @@
 {
   "dedicated_server_interfaces": "Interfaces réseau",
-  "dedicated_server_interfaces_title": "Réseau",
-  "dedicated_server_interfaces_guides": "Guides",
-  "dedicated_server_interfaces_guides_ola": "Comment configurer OLA ?",
-  "dedicated_server_interfaces_introduction": "Vous trouverez dans cette section toutes les informations relatives aux interfaces réseau de votre serveur.",
   "dedicated_server_interfaces_vrack_information": "Veuillez noter que par défaut aucune IP n’existe pour le réseau privé (vRack). En effet vous devez au préalable créer un réseau privé (vRack), y ajouter votre serveur, puis configurez vos IP.",
   "dedicated_server_interfaces_vrack_information_link_label": "En savoir plus sur le vRack.",
 
-  "dedicated_server_interfaces_step_checker_title": "OLA: OVHcloud Link Aggregation",
+  "dedicated_server_interfaces_step_checker_title": "OVHcloud Link Aggregation (OLA)",
   "dedicated_server_interfaces_step_checker_activate": "Activer et configurer OLA en aggrégation privée",
   "dedicated_server_interfaces_step_checker_guides_text": "Pour en savoir plus,",
   "dedicated_server_interfaces_step_checker_guides_link": "consulter nos guides",
